requirements.txt
flask
pdfplumber
python-docx
openpyxl
camelot-py[cv]  # opcional